+++ CHANGELOG.rst
@@ -1,5 +1,18 @@
 What’s new in django-cachalot?
 ==============================
+2.9.0
+-----
+- Add support for Django 6.0 and Python 3.8-3.14 (#277)
+- Fix #280 CachalotPanel not working in Django Debug Toolbar (#277)
+- Fix Recursion Issue in AtomicCache (Issue #274)
+- Allow custom databases (#247)
+
+2.8.0
+-----
+- Add a setting for disabling iterator caching (#263)
+- Add support for testing against the builtin redis backend (#264)
+- Add Django 5.2 support to tox configuration (#272)
+
 2.7.0
 -----
 
@@ -9,7 +22,6 @@ What’s new in django-cachalot?
 -----
 
 - Support for Django 5.1
-
 
 2.6.2
 -----